pyproject.toml
[project]
name = "mlops-airquality"
version = "0.1.0"
description = "MLOps Final Project - Air Quality Prediction with AutoML, MLflow, and FastAPI"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "openpyxl>=3.1.0",
    "h2o>=3.44.0",
    "mlflow>=2.9.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "evidently>=0.4.0",
    "boto3>=1.34.0",
    "psycopg2-binary>=2.9.9",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.1.0",
    "catboost>=1.2.0",
    "python-dotenv>=1.2.1",
    "kaleido==0.2.0",
    "nannyml>=0.13.1",
]
